Adds python3 fixes
This lib is getting old-in-the-tooth. But it cannot be deprecated
until we move all Ansible modules off it.

This adds fixes needed for python3 support.

We needed to change to using suds-jurko instead of fedora suds because
fedora is absolutely not maintained anymore, and jurko at least did us
the honor of having enough python3 support in place that I could fix
bigsuds.